cassandra-lab/backend/requirements.txt
cassandra_driver==3.29.1
fastapi==0.110.0
